use displayorder attribute to order clauses in HTML/DOC; keywords clause to Pres XML: https://github.com/metanorma/metanorma-ogc/issues/567
