+++ lib/isodoc/ieee/html/htmlstyle.scss
@@ -38,10 +38,10 @@ q {
     quotes: none;
 }
 
-blockquote:before,
-blockquote:after,
-q:before,
-q:after {
+blockquote::before,
+blockquote::after,
+q::before,
+q::after {
     content: '';
     content: none;
 }
@@ -228,7 +228,7 @@ ul>li {
     list-style: none;
 }
 
-ul>li>p:first-child:before {
+ul>li>p:first-child::before {
     content: "\2014";
     display: inline-block;
     width: 1em;
@@ -242,7 +242,7 @@ li p {
     line-height: 1.2;
 }
 
-#toc li:before {
+#toc li::before {
     content: " ";
     display: none;
 }
@@ -268,7 +268,7 @@ p.Terms {
     //     box-shadow: inset -5px 0px 10px -5px #1d1d1d !important;
     // }
 
-    li:before {
+    li::before {
         content: " ";
         display: none;
     }
@@ -655,7 +655,7 @@ aside.footnote {
     display: inline;
 }
 
-ol>li>p:before {
+ol>li>p::before {
     content: "";
     display: none;
 }
